FI-39 : migrate from react to nextjs and create base application with get item, collection and post item with static data
