+++ package.json
@@ -14,8 +14,11 @@
   "license": "ISC",
   "dependencies": {
     "isomorphic-unfetch": "^3.0.0",
+    "jquery": "^3.4.1",
     "next": "^9.2.1",
     "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "react-dom": "^16.12.0",
+    "react-markdown": "^4.3.1",
+    "uuidv4": "^6.0.2"
   }
 }
